+++ src/flask/app.py
@@ -235,6 +235,7 @@ class Flask(App):
             "TEMPLATES_AUTO_RELOAD": None,
             "MAX_COOKIE_SIZE": 4093,
             "PROVIDE_AUTOMATIC_OPTIONS": True,
+            "DUPLICATE_URL_RULES": "warn",
         }
     )